Fix accessing command_path as list in templates
In v1 `environment_path` was accessible as a list in templates,
split by the path separator.

In v2 the `environment_path` is known as the `command_path`. This
commit adds support for accessing the `command_path` back in to
Sceptre. For example, the user can now access parts of the
`command_path` in templates like {{ command_path.0 }}.